+++ packages/NumericUpPicker/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rubenpazch/numeric-up-picker
 
+## 3.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @rubenpazch/icons@3.0.4
+
 ## 3.0.3
 
 ### Patch Changes