+++ packages/drawer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rubenpazch/drawer
 
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @rubenpazch/icons@3.0.4
+
 ## 2.0.3
 
 ### Patch Changes